pyproject.toml
[project]
name = "address-label"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.5",
    "platformdirs>=4.5.1",
    "pylabels2>=1.4.4",
]
